+++ README.md
@@ -53,7 +53,7 @@ Bud is an extendable, zero configuration scaffolding CLI for WordPress editor bl
 Bud is currently pre-release. But, you can try it out now:
 
 ```sh
-npx @roots/bud@1.0.0-beta.7 init [project-dir]
+npx @roots/bud@1.0.0-beta.8 init [project-dir]
 ```
 
 This will initialize your project and install `@roots/bud` as a local dev dependency.
