+++ prettier.config.js
@@ -2,7 +2,7 @@ module.exports = {
   arrowParens: 'avoid',
   bracketSpacing: false,
   tabWidth: 2,
-  printWidth: 80,
+  printWidth: 100,
   singleQuote: true,
   jsxBracketSameLine: true,
   useTabs: false,
