+++ src/budfiles/init/templates/.bud/bud.config.json.hbs
@@ -7,9 +7,5 @@
     "author": "{{author}}",
     "email": "{{email}}"
   },
-  "dev": {
-    "protocol": "{{protocol}}",
-    "path": ""
-  },
   "installed": ["bud"]
 }
