+++ src/budfiles/init/templates/.bud/budfiles/example/templates/Component.js.hbs
