--- src/budfiles/init/templates/index.php.bud
@@ -1,10 +1,0 @@
-<?php
-
-require __DIR__ . '/wp/wp-blog-header.php';
-
-print json_encode((object) [
-  'bud' => true,
-  'path' => str_replace(__DIR__ . '/', '', WP_PLUGIN_DIR),
-]);
-
-die();
