+++ src/budfiles/init/templates/package.json.hbs
@@ -16,8 +16,7 @@
   ],
   "private": true,
   "scripts": {
-    "generate": "bud generate new",
-    "generate:list": "bud generate"
+    "generate": "bud generate"
   },
   "devDependencies": {
     "@roots/bud": "roots/bud#master"
